Create src, tests directories
