+++ tests/matcher.test.js
@@ -1,7 +1,8 @@
-const { JEST_DEFAULT_TIMEOUT_MS, DB_SETUP_WAIT_TIME_MS } = require('./constants');
-const DBWrapper = require('./dbwrapper');
-const delay = require('./library');
-const UserPrioritizer = require('./userprioritizer');
+const { JEST_DEFAULT_TIMEOUT_MS, DB_SETUP_WAIT_TIME_MS } = require('../src/constants');
+const DBWrapper = require('../src/dbwrapper');
+const delay = require('../src/utility');
+const Matcher = require('../src/matcher');
+const UserPrioritizer = require('../src/userprioritizer');
 
 const dbWrapper = new DBWrapper();
 const userPrioritizer = new UserPrioritizer();
@@ -99,9 +100,10 @@ afterAll(async () => {
   await delay(DB_SETUP_WAIT_TIME_MS);
 }, JEST_DEFAULT_TIMEOUT_MS + DB_SETUP_WAIT_TIME_MS);
 
-test('user with no previous meeting should be on top of priority list', async () => {
+test('xxxxxxxs', async () => {
   const dec31st2021DateString = new Date(2021, 12, 31).toISOString();
   const dec21st2021DateString = new Date(2021, 12, 21).toISOString();
+
   const batchWriteParamsForPutting = {
     RequestItems: {
       devcatchupmeetings: [
@@ -109,9 +111,29 @@ test('user with no previous meeting should be on top of priority list', async ()
           PutRequest: {
             Item: {
               KEY: { S: 'email1' },
-              email1: { S: 'sakibulmowla@gmail.com' },
+              email1: { S: 'masum.nayeem@gmail.com' },
               email2: { S: 'biswajit.sust@gmail.com' },
               timestamp: { S: dec31st2021DateString },
+            },
+          },
+        },
+        {
+          PutRequest: {
+            Item: {
+              KEY: { S: 'email1' },
+              email1: { S: 'biswajit.sust@gmail.com' },
+              email2: { S: 'masum.nayeem@gmail.com' },
+              timestamp: { S: dec31st2021DateString },
+            },
+          },
+        },
+        {
+          PutRequest: {
+            Item: {
+              KEY: { S: 'email1' },
+              email1: { S: 'sakibulmowla@gmail.com' },
+              email2: { S: 'biswajit.sust@gmail.com' },
+              timestamp: { S: dec21st2021DateString },
             },
           },
         },
@@ -131,46 +153,41 @@ test('user with no previous meeting should be on top of priority list', async ()
   await dbWrapper.batchWriteItem(batchWriteParamsForPutting);
 
   const orderedUserList = await userPrioritizer.getOrderedListOfUsers('dev');
+  const matcher = new Matcher(orderedUserList);
 
-  expect(orderedUserList.length).toBe(3);
-  expect(orderedUserList[0].email).toBe('masum.nayeem@gmail.com');
-});
-
-test('users should be prioritized in latemost last meeting order', async () => {
-  const dec22nd2021DateString = new Date(2021, 12, 22).toISOString();
-  const dec11th2021DateString = new Date(2021, 12, 11).toISOString();
-  const batchWriteParamsForPutting = {
-    RequestItems: {
-      devcatchupmeetings: [
-        {
-          PutRequest: {
-            Item: {
-              KEY: { S: 'email1' },
-              email1: { S: 'masum.nayeem@gmail.com' },
-              email2: { S: 'sakibulmowla@gmail.com' },
-              timestamp: { S: dec22nd2021DateString },
-            },
-          },
-        },
-        {
-          PutRequest: {
-            Item: {
-              KEY: { S: 'email1' },
-              email1: { S: 'sakibulmowla@gmail.com' },
-              email2: { S: 'biswajit.sust@gmail.com' },
-              timestamp: { S: dec11th2021DateString },
-            },
-          },
-        },
-      ],
+  const preferenceListOfSakibul = await matcher.getPreferenceList(
+    {
+      firstname: 'Sakibul',
+      lastname: 'Mowla',
+      email: 'sakibulmowla@gmail.com',
+      timestamp: dec21st2021DateString,
     },
-  };
-  await dbWrapper.batchWriteItem(batchWriteParamsForPutting);
+    'dev',
+  );
+  expect(preferenceListOfSakibul).toStrictEqual([
+    'masum.nayeem@gmail.com',
+    'biswajit.sust@gmail.com',
+  ]);
 
-  const orderedUserList = await userPrioritizer.getOrderedListOfUsers('dev');
+  const preferenceListOfBiswajit = await matcher.getPreferenceList(
+    {
+      firstname: 'Biswajit',
+      lastname: 'Debnath',
+      email: 'biswajit.sust@gmail.com',
+      timestamp: dec31st2021DateString,
+    },
+    'dev',
+  );
+  expect(preferenceListOfBiswajit).toStrictEqual([
+    'sakibulmowla@gmail.com',
+    'masum.nayeem@gmail.com',
+  ]);
 
-  expect(orderedUserList.length).toBe(3);
-  expect(orderedUserList[0].email).toBe('biswajit.sust@gmail.com');
-  expect(orderedUserList[1].email).toBe('masum.nayeem@gmail.com');
-  expect(orderedUserList[2].email).toBe('sakibulmowla@gmail.com');
+  const matchedGroups = await matcher.getMatchedGroups('dev');
+  expect(matchedGroups).toStrictEqual([
+    [
+      'sakibulmowla@gmail.com',
+      'masum.nayeem@gmail.com',
+    ],
+  ]);
 });
